+++ src/dll/fuse/fuse.c
@@ -106,6 +106,7 @@ static struct fuse_opt fsp_fuse_core_opts[] =
     FSP_FUSE_CORE_OPT("FlushOnCleanup=", set_FlushOnCleanup, 1),
     FSP_FUSE_CORE_OPT("LegacyUnlinkRename=", set_LegacyUnlinkRename, 1),
     FSP_FUSE_CORE_OPT("AllowRelSymlinksAcrossFileSystem", set_AllowRelSymlinksAcrossFileSystem, 1),
+    FSP_FUSE_CORE_OPT("WslFeatures", set_WslFeatures, 1),
     FSP_FUSE_CORE_OPT("ThreadCount=%u", ThreadCount, 0),
     FUSE_OPT_KEY("UNC=", 'U'),
     FUSE_OPT_KEY("--UNC=", 'U'),
@@ -673,6 +674,7 @@ static int fsp_fuse_core_opt_proc(void *opt_data0, const char *arg, int key,
             "    -o LegacyUnlinkRename      do not support new POSIX unlink/rename\n"
             "    -o AllowRelSymlinksAcrossFileSystem\n"
             "                                allow relative symlinks to cross file systems\n"
+            "    -o WslFeatures             enable WSL drvfs metadata queries\n"
             "    -o ThreadCount             number of file system dispatcher threads\n"
             "    -o UserOwner               use caller user as file owner\n"
             "    -o uidmap=UID:SID[;...]    explicit UID <-> SID map (max 8 entries)\n"
@@ -900,6 +902,8 @@ FSP_FUSE_API struct fuse *fsp_fuse_new(struct fsp_fuse_env *env,
         opt_data.VolumeParams.SupportsPosixUnlinkRename = FALSE;
     if (opt_data.set_AllowRelSymlinksAcrossFileSystem)
         opt_data.VolumeParams.AllowRelSymlinksAcrossFileSystem = TRUE;
+    if (opt_data.set_WslFeatures)
+        opt_data.VolumeParams.WslFeatures = TRUE;
     if (FIELD_OFFSET(struct fuse_operations, link) + sizeof ops->link <= opsize &&
         0 != ops->link)
         opt_data.VolumeParams.HardLinks = TRUE;