+++ src/dll/fuse/library.h
@@ -258,7 +258,8 @@ struct fsp_fuse_core_opt_data
         set_KeepFileCache,
         set_FlushOnCleanup,
         set_LegacyUnlinkRename,
-        set_AllowRelSymlinksAcrossFileSystem;
+        set_AllowRelSymlinksAcrossFileSystem,
+        set_WslFeatures;
     unsigned ThreadCount;
     FSP_FSCTL_VOLUME_PARAMS VolumeParams;
     UINT16 VolumeLabelLength;